Unify exchange_rate to decimal(16,6) and fix truncation of small rates
shop_order.exchange_rate was decimal(15,2): a small snapshot rate (a
large currency priced against a small-unit base, e.g. a VND base plus
USD ~ 0.00004) truncated to 0.00, so the order lost its rate for good.
The three columns were also inconsistent (shop_currency and
shop_order_detail were float).

Unify exchange_rate to decimal(16,6) (exact to 0.000001) on shop_currency,
shop_order and shop_order_detail. Ship it two ways for the open-source
multi-environment audience: the create-tables migration for fresh installs,
plus an idempotent ALTER upgrade migration (guards on NUMERIC_SCALE, no-op
when already correct, empty data-preserving down()) wired into
gp247:shop-update, runnable on shared hosting (no cron/queue).

Cast exchange_rate to float on the three models: Laravel returns a decimal
column as a string, but the rate is multiplied against money, so consumers
need a number while the DB keeps exact precision.

Fix the input layer too: the admin currency form used
step="0.0001", which the browser rejects for non-multiples (e.g. 0.00004);
switch to step="any" so small rates can be typed (server validation
numeric,gt:0 unchanged; decimal(16,6) rounds at 6 places).

BREAKING: changing an existing column type is a deliberate exception to
NFR-MAINT-002 (a data-loss fix, and there is no 1.x->2.0 in-place upgrade).
No backfill: already-truncated 0.00 rates are unrecoverable and inferring
from the current rate would be wrong (the rate is a point-in-time snapshot).

US-CMP-exchange-rate-precision (Nhom B). See ADR
compat-foundation_exchange-rate-precision.

+++ src/Admin/Database/Migrations/00_00_00_create_tables_shop.php
@@ -64,7 +64,12 @@ return new class extends Migration
                 $table->string('name', 255);
                 $table->string('code', 10)->unique();
                 $table->string('symbol', 10);
-                $table->float('exchange_rate');
+                // WHY: decimal(16,6) unifies exchange_rate across all three tables
+                // (this, shop_order, shop_order_detail) and keeps rates as small as
+                // ~0.00004 exactly — a large currency priced against a small-unit base
+                // (e.g. VND base). float was imprecise; shop_order was decimal(15,2)
+                // which truncated such rates to 0.00 (ADR compat-foundation_exchange-rate-precision).
+                $table->decimal('exchange_rate',16,6);
                 $table->tinyInteger('precision')->default(2);
                 $table->tinyInteger('symbol_first')->default(0);
                 $table->string('thousands')->default(',');
@@ -90,7 +95,11 @@ return new class extends Migration
                 $table->decimal('other_fee',15,2)->nullable()->default(0);
                 $table->decimal('total',15,2)->nullable()->default(0);
                 $table->string('currency', 10);
-                $table->decimal('exchange_rate',15,2)->nullable();
+                // WHY: decimal(16,6), not the former decimal(15,2), so a small
+                // snapshot rate (large currency vs small-unit base, ~0.00004) is not
+                // truncated to 0.00 — the order would otherwise lose its exchange rate
+                // permanently (RISK-TECH-exchange-rate-truncation, ADR compat-foundation_exchange-rate-precision).
+                $table->decimal('exchange_rate',16,6)->nullable();
                 $table->decimal('received',15,2)->nullable()->default(0);
                 $table->decimal('balance',15,2)->nullable()->default(0);
                 $table->string('first_name', 100);
@@ -135,7 +144,10 @@ return new class extends Migration
                 $table->decimal('tax',15,2)->default(0);
                 $table->string('sku', 50);
                 $table->string('currency', 10);
-                $table->float('exchange_rate')->nullable();
+                // WHY: decimal(16,6) to match shop_order / shop_currency — one exact
+                // type for exchange_rate everywhere (NFR-MAINT-exchange-rate-type-consistency,
+                // ADR compat-foundation_exchange-rate-precision).
+                $table->decimal('exchange_rate',16,6)->nullable();
                 // WHY: stores json_encode(options) whose length grows with the number of
                 // attribute groups / long names — a fixed varchar(100) truncates it
                 // (RISK-TECH-order-attribute-truncation, ADR storefront_attribute-price-integrity).
